Refactor project into modular structure and fix naming issues
## Major Refactoring
- Split monolithic 3,434-line client.rs into organized modules:
  - client/ - API operations, authentication, downloads (3 files)
  - naming/ - Name generation and templates (3 files)
  - models/ - Data structures (3 files)
  - config/ - Path management (1 file)
  - utils/ - Output formatting and errors (2 files)
- Created lib.rs as library root with proper exports
- Updated main.rs to use new library structure
- All files now under 500 lines for better maintainability

## Naming System Improvements
- Fix thread pitch extraction for imperial threads (5/16"-18 → 5/16"x18)
- Fix screw size abbreviation for washers (No. 6 → 6)
- Standardize aluminum abbreviation across all fastener types (Al → AL)
- Enhanced abbreviate_value() function with special handling for:
  - Thread sizes (preserve full designation)
  - Screw sizes (extract number from "No. X" format)

## Testing Verified
- ✅ Thread pitch: LN-SS188-5/16x18 (locknut 91831A030)
- ✅ Screw sizes: FW-SS188-6 (washer 92141A008)
- ✅ Aluminum: FSO-AL-4X40-3.5 (standoff 91780A053)
- ✅ All existing functionality preserved
